migrated dish_name_tag_average
