+++ server/backend/migration/20231015000000.go
@@ -69,13 +69,29 @@ func (n *CafeteriaRatingTagsAverage) TableName() string {
 	return "cafeteria_rating_tag_average"
 }
 
+// DishNameTagAverage stores all precomputed values for the DishName ratings
+type DishNameTagAverage struct {
+	DishNameTagAverage int64   `gorm:"primary_key;AUTO_INCREMENT;column:dishNameTagAverage;type:int;not null;" json:"dishNameTagAverage" `
+	CafeteriaID        int64   `gorm:"column:cafeteriaID;foreignKey:cafeteria;type:int;not null;" json:"cafeteriaID"`
+	TagID              int64   `gorm:"column:tagID;foreignKey:DishNameTagOption;type:int;not null;" json:"tagID"`
+	Average            float32 `gorm:"column:average;type:float;not null;" json:"average" `
+	Min                int8    `gorm:"column:min;type:int;not null;" json:"min"`
+	Max                int8    `gorm:"column:max;type:int;not null;" json:"max"`
+	Std                float32 `gorm:"column:std;type:float;not null;" json:"std"`
+}
+
+// TableName sets the insert table name for this struct type
+func (n *DishNameTagAverage) TableName() string {
+	return "dish_name_tag_average"
+}
+
 // migrate20231015000000
 // migrates the static data for the canteen rating system and adds the necessary cronjob entries
 func (m TumDBMigrator) migrate20231015000000() *gormigrate.Migration {
 	return &gormigrate.Migration{
 		ID: "20231015000000",
 		Migrate: func(tx *gorm.DB) error {
-			tables := []string{"cafeteria_rating_average", "dish_rating_average", "dish_rating_tag_average", "cafeteria_rating_tag_average"}
+			tables := []string{"cafeteria_rating_average", "dish_rating_average", "dish_rating_tag_average", "cafeteria_rating_tag_average", "dish_name_tag_average"}
 			for _, table := range tables {
 				if err := tx.Migrator().DropTable(table); err != nil {
 					return err
@@ -109,16 +125,23 @@ JOIN cafeteria_rating_tag crt ON cr.cafeteriaRating = crt.correspondingRating
 GROUP BY cr.cafeteriaID, crt.tagID`).Error; err != nil {
 				return err
 			}
+			if err := tx.Exec(`CREATE VIEW dish_name_tag_statistics AS
+SELECT mr.cafeteriaID as cafeteriaID, mnt.tagnameID as tagID, AVG(mnt.points) as average, MAX(mnt.points) as max, MIN(mnt.points) as min, STD(mnt.points) as std
+FROM dish_rating mr
+JOIN dish_name_tag mnt ON mr.dishRating = mnt.correspondingRating
+GROUP BY mr.cafeteriaID, mnt.tagnameID`).Error; err != nil {
+				return err
+			}
 			return nil
 		},
 		Rollback: func(tx *gorm.DB) error {
-			createdViews := []string{"cafeteria_rating_statistics", "dish_rating_statistics", "dish_rating_tag_statistics", "cafeteria_rating_tag_average"}
+			createdViews := []string{"cafeteria_rating_statistics", "dish_rating_statistics", "dish_rating_tag_statistics", "cafeteria_rating_tag_statistics", "dish_name_tag_statistics"}
 			for _, view := range createdViews {
 				if err := tx.Exec("DROP VIEW IF EXISTS " + view).Error; err != nil {
 					return err
 				}
 			}
-			return tx.AutoMigrate(&CafeteriaRatingAverage{}, &DishRatingAverage{}, &DishRatingTagAverage{}, &CafeteriaRatingTagsAverage{})
+			return tx.AutoMigrate(&CafeteriaRatingAverage{}, &DishRatingAverage{}, &DishRatingTagAverage{}, &CafeteriaRatingTagsAverage{}, &DishNameTagAverage{})
 		},
 	}
 }
